Implemented more V2 api

+++ IssueTracker.API/Security/ISecurityService.cs
@@ -9,13 +9,13 @@ namespace IssueTracker.API.Security
         bool AddPermission(UserEntity user, string permission);
         bool RemovePermission(UserEntity user, string permission);
 
-        UserEntity GetUser(int id);
+        UserEntity GetUser(long id);
         UserEntity GetCurrentUser();
 
         bool TryAuthenticate(string username, string password, out UserEntity user);
 
-        AuthTokenEntity IssueToken(int userId);
-        AuthTokenEntity IssueToken(int userId, TimeSpan expiresIn);
+        AuthTokenEntity IssueToken(long userId);
+        AuthTokenEntity IssueToken(long userId, TimeSpan expiresIn);
         bool TryValidateToken(string token, out UserEntity user);
 
         void ResetPassword(UserEntity user, string password);
